pyproject.toml
[project]
name = "pynotex"
version = "1.0.0"
description = "A privacy-first, open-source alternative to NotebookLM - Python implementation"
readme = "README.md"
requires-python = ">=3.12.2"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "knowledge-management", "rag", "notebook", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Web 框架
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.12",

    # LLM
    "openai==1.54.0",
    "httpx>=0.27.0,<0.28.0",
    "google-generativeai==0.8.3",

    # 文档处理
    "markitdown==0.0.1a3",
    "pymupdf==1.24.13",
    "python-docx==1.1.2",
    "python-pptx==1.0.2",
    "openpyxl==3.1.5",
    "beautifulsoup4==4.12.3",

    # 数据库
    "aiosqlite==0.20.0",

    # 工具
    "pydantic==2.10.0",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
